Move cmdparser.*, command.*, and execution.* from celengine to celestia, modify celephem to OBJECT library and merge the objects in the celengine library, fix Spice dependency
