+++ CMakeLists.txt
@@ -109,8 +109,9 @@ if(ENABLE_SPICE)
   if(NOT CSPICE_FOUND)
     message(STATUS "Using cspice submodule")
     add_subdirectory("${CMAKE_SOURCE_DIR}/thirdparty/Spice")
-
-    include_directories("${CMAKE_SOURCE_DIR}/thirdparty/Spice/include")
+    get_target_property(SPICE_INCLUDE_DIR Spice INCLUDE_DIRECTORIES)
+    include_directories(${SPICE_INCLUDE_DIR})
+    message(STATUS "Spice include directories: ${SPICE_INCLUDE_DIR}")
     add_library(CSPICE::CSPICE ALIAS Spice)
   endif()
   add_definitions(-DUSE_SPICE)
