+++ src/CMakeLists.txt
@@ -1,7 +1,6 @@
 set(CELESTIA_LIBS
   cel3ds
   celengine
-  celephem
   celestia
   celmath
   celmodel
@@ -12,6 +11,9 @@ set(CELESTIA_LIBS
 foreach(lib ${CELESTIA_LIBS})
   add_subdirectory(${lib})
 endforeach()
+
+# Compiled celephem objects are merged with the celengine library
+add_subdirectory(celephem)
 
 if (ENABLE_TOOLS)
   add_subdirectory(tools)
