+++ src/celmath/CMakeLists.txt
@@ -14,7 +14,7 @@ set(CELMATH_SOURCES
 )
 
 add_library(celmath STATIC ${CELMATH_SOURCES})
-cotire(celmath)
+
 #[[
 add_library(celmath SHARED ${CELMATH_SOURCES})
 set_target_properties(celmath PROPERTIES
@@ -22,4 +22,7 @@ set_target_properties(celmath PROPERTIES
   SOVERSION ${VERSION}
 )
 ]]#
+
+cotire(celmath)
+
 #install(TARGETS celmath LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR})
