+++ src/celutil/CMakeLists.txt
@@ -38,7 +38,7 @@ else()
 endif()
 
 add_library(celutil STATIC ${CELUTIL_SOURCES})
-cotire(celutil)
+
 #[[
 add_library(celutil SHARED ${CELUTIL_SOURCES})
 set_target_properties(celutil PROPERTIES
@@ -46,4 +46,10 @@ set_target_properties(celutil PROPERTIES
   SOVERSION ${VERSION}
 )
 ]]#
+
+add_dependencies(celutil celmath)
+target_link_libraries(celutil celmath)
+
+cotire(celutil)
+
 #install(TARGETS celutil LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR})
